+++ plugin/src/main/kotlin/com/moengage/gradle/android/library/plugin/utils/Constants.kt
@@ -3,11 +3,11 @@ package com.moengage.gradle.android.library.plugin.utils
 
 // Build Versions
 internal const val ANDROID_BUILD_CONFIG_MINIMUM_SDK_VERSION = 23
-internal const val ANDROID_BUILD_CONFIG_COMPILE_SDK_VERSION = 36
-internal const val ANDROID_BUILD_CONFIG_TARGET_SDK_VERSION = 36
+internal const val ANDROID_BUILD_CONFIG_COMPILE_SDK_VERSION = 35
+internal const val ANDROID_BUILD_CONFIG_TARGET_SDK_VERSION = 35
 
 // Library Versions
-internal const val JACOCO_VERSION = "0.8.8"
+internal const val JACOCO_VERSION = "0.8.14"
 
 // Plugin IDs
 internal const val PLUGIN_ID_ANDROID_LIBRARY = "com.android.library"